Harrier-project/requirements.txt
requests
python-dotenv
sqlalchemy
pandas
mysql-connector-python
